Fix RTSP timestamp bugs causing buffering
Three fixes for broken RTSP PTS that caused players to buffer:

1. IMPSystem.hpp: Fix int32 overflow in RebaseTimeStamp computation.
   On MIPS32, tv_sec * 1000000 overflows when uptime > 35.8 minutes,
   corrupting all IMP encoder timestamps. Cast to uint64_t before
   multiplying.

2. IMPDeviceSource.cpp: Fix min_audio_step_us for non-AAC codecs.
   Was computing single-sample duration (21us for 48kHz) instead of
   frame duration (40000us for Opus 1920-sample frames). The too-small
   step caused the monotonicity guard to produce broken audio PTS.

3. IMPDeviceSource.cpp: Re-anchor on large forward PTS jumps.
   The audio/video guards treated both backwards and large forward
   jumps identically, clamping to lastPts + minStep. On the second
   RTSP connection, the gap between sessions triggered the guard on
   every frame, collapsing PTS to near-zero increments. Now backwards
   jumps enforce monotonicity while forward jumps re-anchor to the
   new timestamp.

+++ src/IMPSystem.hpp
@@ -25,7 +25,8 @@ public:
     struct timespec timeSinceBoot;
     clock_gettime(CLOCK_MONOTONIC, &timeSinceBoot);
 
-    uint64_t imp_time_base = (timeSinceBoot.tv_sec * 1000000) + (timeSinceBoot.tv_nsec / 1000);
+    uint64_t imp_time_base = (static_cast<uint64_t>(timeSinceBoot.tv_sec) * 1000000ULL)
+                           + (static_cast<uint64_t>(timeSinceBoot.tv_nsec) / 1000ULL);
     IMP_System_RebaseTimeStamp(imp_time_base);
 
     LOG_DEBUG("IMP_System_RebaseTimeStamp(" << imp_time_base << ");");
